moved replication related services and controllers into module
