+++ src/replication/couch-proxy/couchdb-dtos/all-docs.dto.ts
