+++ src/replication/couch-proxy/couchdb-dtos/bulk-docs.dto.ts
