+++ src/replication/couch-proxy/couchdb-dtos/bulk-get.dto.ts
