+++ src/replication/couch-proxy/couchdb-dtos/revision-diff.dto.ts
