+++ src/replication/document-filter/access-control-entry.ts
